backend/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
Flask_SocketIO==5.4.1
matplotlib==3.9.4
numpy==2.2.0
opencv_python==4.10.0.84
Pillow==11.0.0
psycopg2==2.9.10
Requests==2.32.3
torch==2.5.1
torchvision==0.20.1
